+++ Assets/Scripts/Vanilla/GameContent/RandomChinaEvents/Tarots/TheTowerEvent.cs
@@ -30,6 +30,6 @@ namespace MVZ2.GameContent.RandomChinaEvents
             }
         }
         [TranslateMsg("随机瓷器事件名称", VanillaStrings.CONTEXT_RANDOM_CHINA_EVENT_NAME)]
-        public const string NAME = "塔-XVI";
+        public const string NAME = "XVI-塔";
     }
 }